\name{FieldSpectroscopyCC-package}
\alias{FieldSpectroscopyCC-package}
\alias{FieldSpectroscopyCC}
\docType{package}
\title{R package for Characterization and Calibration of spectrometers}
}
\description{R package for Characterization and Calibration of spectrometers}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FieldSpectroscopyCC}
\packageIndices{FieldSpectroscopyCC}
}
\author{Tommaso Julitta, Mirco Migliavacca, Thomas Wutzler}

Maintainer: \packageMaintainer{FieldSpectroscopyCC}
}

\keyword{ package }


